changed access from magic asseccors to setter and getter

+++ application/models/Resident.php
@@ -63,6 +63,11 @@ class Resident extends BaseModel {
 	}
 
 	public function delete(){
+		
+		if ($this->getCashAccount()->getBalanceInCent() != 0) {
+			throw new Exception_BuisnessLogic("Das Konto von ".$this->getName()." ist nicht ausgeglichen.");
+		}
+		
 		if ($this->getCashAccount()->delete() > 0){
 			$this->delete();
 		}
